Refactor tournament registration management interfaces
- Added `ITournamentRegistrationsView` interface for managing tournament registrations.
- Introduced `TournamentRegistrationsPresenter` class to handle presentation logic, including methods for retrieval, deletion, and updates.
- Modified `TournamentRegistrationViewModel` to implement `ITournamentRegistrationViewModel` and updated its constructor.
- Added XML documentation for clarity and maintainability.
- Removed old interfaces and classes, replacing them with the new structured implementations.
