--- source/main/Registrations/Retrieve/RetrieveTournamentRegistrationsPresenter.cs
@@ -1,133 +1,0 @@
-﻿using System.Globalization;
-using Microsoft.Extensions.DependencyInjection;
-
-namespace BowlingMegabucks.TournamentManager.Registrations.Retrieve;
-internal class TournamentRegistrationsPresenter
-{
-    private readonly ITournamentRegistrationsView _view;
-
-    private readonly IAdapter _registrationsAdapter;
-    private readonly Squads.Retrieve.IAdapter _squadsAdapter;
-    private readonly Sweepers.Retrieve.IAdapter _sweepersAdapter;
-
-    private readonly Lazy<Delete.IAdapter> _deleteAdapter;
-    private Delete.IAdapter DeleteAdapter => _deleteAdapter.Value;
-
-    private readonly Lazy<Update.IAdapter> _updateAdapter;
-    private Update.IAdapter UpdateAdapter => _updateAdapter.Value;
-
-    public TournamentRegistrationsPresenter(ITournamentRegistrationsView view, IServiceProvider services)
-    {
-        _view = view;
-
-        _registrationsAdapter = services.GetRequiredService<IAdapter>();
-        _squadsAdapter = services.GetRequiredService<Squads.Retrieve.IAdapter>();
-        _sweepersAdapter = services.GetRequiredService<Sweepers.Retrieve.IAdapter>();
-
-        _deleteAdapter = new Lazy<Delete.IAdapter>(services.GetRequiredService<Delete.IAdapter>);
-        _updateAdapter = new Lazy<Update.IAdapter>(services.GetRequiredService<Update.IAdapter>);
-    }
-
-    /// <summary>
-    /// Unit Test Constructor
-    /// </summary>
-    /// <param name="mockView"></param>
-    /// <param name="mockRegistrationsAdapter"></param>
-    /// <param name="mockSquadsAdapter"></param>
-    /// <param name="mockSweepersAdapter"></param>
-    /// <param name="mockDeleteAdapter"></param>
-    /// <param name="mockUpdateAdapter"></param>
-    internal TournamentRegistrationsPresenter(ITournamentRegistrationsView mockView, IAdapter mockRegistrationsAdapter, Squads.Retrieve.IAdapter mockSquadsAdapter, Sweepers.Retrieve.IAdapter mockSweepersAdapter, Delete.IAdapter mockDeleteAdapter, Update.IAdapter mockUpdateAdapter)
-    {
-        _view = mockView;
-        _registrationsAdapter = mockRegistrationsAdapter;
-        _squadsAdapter = mockSquadsAdapter;
-        _sweepersAdapter = mockSweepersAdapter;
-        _deleteAdapter = new Lazy<Delete.IAdapter>(() => mockDeleteAdapter);
-        _updateAdapter = new Lazy<Update.IAdapter>(() => mockUpdateAdapter);
-    }
-
-    public async Task ExecuteAsync(CancellationToken cancellationToken)
-    {
-        var registrations = await _registrationsAdapter.ExecuteAsync(_view.TournamentId, cancellationToken).ConfigureAwait(true);
-        var squads = await _squadsAdapter.ExecuteAsync(_view.TournamentId, cancellationToken).ConfigureAwait(true);
-        var sweepers = await _sweepersAdapter.ExecuteAsync(_view.TournamentId, cancellationToken).ConfigureAwait(true);
-
-        var errors = new[] { _registrationsAdapter.Error, _squadsAdapter.Error, _sweepersAdapter.Error }.Where(error => error != null).ToList();
-
-        if (errors.Count > 0)
-        {
-            _view.DisplayError(string.Join(Environment.NewLine, errors.Select(error => error!.Message).Distinct()));
-
-            return;
-        }
-
-        const string dateFormat = "MM/dd/yy hh:mmtt";
-        var squadsDictionary = squads.ToDictionary(squad => squad.Id, squad => squad.Date.ToString(dateFormat, CultureInfo.CurrentCulture));
-        var squadEntries = squadsDictionary.ToDictionary(squad => squad.Value, squad => registrations.Count(registration => registration.SquadsEntered.Contains(squad.Key)));
-        _view.SetSquadEntries(squadEntries);
-        _view.BindSquadDates(squadsDictionary);
-
-        var sweepersDictionary = sweepers.ToDictionary(sweeper => sweeper.Id, sweeper => sweeper.Date.ToString(dateFormat, CultureInfo.CurrentCulture));
-        var sweeperEntries = sweepersDictionary.ToDictionary(sweeper => sweeper.Value, sweeper => registrations.Count(registration => registration.SweepersEntered.Contains(sweeper.Key)));
-        sweeperEntries.Add("Super Sweeper", registrations.Count(registration => registration.SuperSweeperEntered));
-        _view.SetSweeperEntries(sweeperEntries);
-        _view.BindSquadDates(sweepersDictionary);
-
-        _view.BindRegistrations(registrations.OrderBy(registration => registration.LastName).ThenBy(registration => registration.FirstName));
-
-        var divisionEntries = registrations.GroupBy(registration => registration.DivisionName).ToDictionary(g => g.Key, g => g.Sum(r => r.SquadsEnteredCount));
-        _view.SetDivisionEntries(divisionEntries);
-    }
-
-    public async Task DeleteAsync(RegistrationId id, CancellationToken cancellationToken)
-    {
-        if (!_view.Confirm("Are you sure you want to delete this bowler's entire registration?"))
-        {
-            return;
-        }
-
-        await DeleteAdapter.ExecuteAsync(id, cancellationToken).ConfigureAwait(true);
-
-        if (DeleteAdapter.Error != null)
-        {
-            _view.DisplayError(DeleteAdapter.Error.Message);
-        }
-        else
-        {
-            _view.RemoveRegistration(id);
-        }
-    }
-
-    public void UpdateBowlerName(BowlerId id)
-    {
-        var updatedName = _view.UpdateBowlerName(id);
-
-        if (updatedName == null)
-        {
-            return;
-        }
-
-        _view.UpdateBowlerName(updatedName!);
-    }
-
-    public async Task AddSuperSweeperAsync(RegistrationId id, CancellationToken cancellationToken)
-    {
-        if (!_view.Confirm("Are you sure you want to add a Super Sweeper entry to this registration?"))
-        {
-            return;
-        }
-
-        await UpdateAdapter.AddSuperSweeperAsync(id, cancellationToken).ConfigureAwait(false);
-
-        if (UpdateAdapter.Errors.Any())
-        {
-            _view.DisplayError(UpdateAdapter.Errors.First().Message);
-
-            return;
-        }
-
-        _view.DisplayMessage("Super Sweeper has been added");
-        _view.UpdateBowlerSuperSweeper(id);
-    }
-}
